// src/app/Forms/models/form-transaction.model.ts
export class formTransaction {
    transaction: string = "";
    operator: string = "";
    where: string = "";
    destination: string = "";
    date: string = "";
    manufacturer: string = "";
    product: string = "";
    //quantities
    full: number = 0;
    tQ: number = 0;
    half: number = 0;
    part: number = 0;
}